Update dp_fork_batch with last commits (bwd integration, US76 model).
For a (perhaps not too distante) future: the code is plainly redundant at this stage. It is not necessary to replicate the whole code from dp/ in the dp_fork_batch/ fork as to perform batch propagations; in fact only the main program and some other sources (settings, io) are affected. The structure of the program should thus be reformed.