Add script for propagation batches in tolerance.
The Python 2.7 script batch_tol.py has been created to perform propagation batches along tolerance values. It takes as an argument a "master" directory in which to save the output from each of the propagations, and it saves the integration statistics of the batch propagation in a 'summary.csv' file thereby contained.

The stat path has been removed from Thalassa; the stat file is now saved in the only output directory, i.e. 'outpath'. The separator has been set to whitespace in the stat file.