Add switch architecture in the propagation loop.
The propagation loop has been modified to accommodate the switch of coordinate systems. A new module, COORD_SYST, has been created to contain the related procedures.

The input format has been changed from orbital elements to Cartesian coordinates, in prevision of the numerical experiments on periodic orbits.

New "ephemerides" for the Moon have been created, in which it is considered as travelling on a Keplerian orbit around the Earth. The gravitational parameter has also been modified in accordance to the one reported in Szebehely (1967), p. 510.