Fix current output directory.
thalassa_main was copying the input files in in/*.* in the output directory. This caused subsequent propagations to use wrong inputs. This has now been fixed.

In batch_tol.py, the output folder (which in general is '$MASTER_PATH/tol*/) is cleared prior to each  iteration in tolerance.
Also, Thalassa now appends lines to the statistics file.