Bugfix reading of gravitational coefficients.
The read loop on the gravitational coefficient matrices Clm, Slm
in phys_const.f90 was performed incorrectly. This caused a wrong
initialization of Clm, Slm any time that gord < gdeg. The loop
has been rewritten in a simpler way, and the remaining modules
using Clm, Slm have been modified for consistency.

This is a conflict-resolving commit cherry-picked from the
master branch.