Remove "model" input flag.
The  "model" input flag was originally introduced to switch
between two physical models that reflected those of the STELA and
SWIFT-SAT propagators. However, in pretty much all of the versions
the only impact of this flag has been to slightly change the computation
of the GMST. I decided to remove the flag since this has never been actually
used
in months.